Restructured sequelize part
